% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus.R
\name{getScopusMetaData}
\alias{getScopusMetaData}
\title{getScopusMetaData}
\usage{
getScopusMetaData(searchResults, myAPIKey, ordinationFunction = FALSE)
}
\arguments{
\item{searchResults}{a Data Frame produced by the functions
\code{\link[ginko]{searchScopus}} or
\code{\link[ginko]{createTextMatrixFromPDF}}}

\item{myAPIKey}{Your private Elsevier API key for the server communication.
You can request one at \url{https://dev.elsevier.com/user/login}}

\item{ordinationFunction}{FALSE by default}
}
\value{
A data frame with all meta data of the articles provided to the function
}
\description{
takes a Data Frame of DOI numbers and/or Scopus-IDs (like
    \code{\link[ginko]{searchScopus}}) and downloads the abstracts and
    titles matching those numbers and returning a matrix containing both.
}
\examples{
\dontrun{
    processedMetaMatrix$MetaMatrix <- getScopusMetaData(processedMetaMatrix$MetaMatrix, myAPIKey)
    }
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
