% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{processMetaDataMatrix}
\alias{processMetaDataMatrix}
\title{processMetaDataMatrix}
\usage{
processMetaDataMatrix(metaMatrix, language = "SMART", stemWords = TRUE,
  ignoreWords = c(), useMoritz = FALSE, moritzFile = "moritz.csv",
  saveToWd = TRUE, ordinationFunction = FALSE, longMessages = FALSE)
}
\arguments{
\item{metaMatrix}{metaMatrix created through \code{\link[ginko]{getScopusMetaData}}
or \code{\link[ginko]{createTextMatrixFromPDF}}.}

\item{language}{this defines the stopwords to be filtered. the default is
"english". Look at \code{\link[tm]{stopwords}} for more information.}

\item{stemWords}{can be \code{TRUE} of \code{FALSE}. Transforms every word
to its stem, so variants of the same words are treated equally. Look
at \code{\link[tm]{stemDocument}} for more information.}

\item{ignoreWords}{a vector of words to be ignored.}

\item{useMoritz}{allows to specify the exact words that will be kept. Should
only be used with caution. The list can be specified in \code{moritzFile}.}

\item{moritzFile}{the list specifying which exact words will be kept.
\code{useMoritz} has to be set to \code{TRUE}}

\item{saveToWd}{a logical parameter whether or not to save the output of the
function to the working directory. This is especially useful for later
analysis steps and can be read in by using \code{\link[base]{saveRDS}}.}

\item{ordinationFunction}{placeholder}

\item{longMessages}{placeholder}
}
\value{
returns a list with entry one being a binary matrix with all occuring
    words as columns and all papers as rows. The second list entry is an
    updated version of the input \code{metaMatrix}, excluding entries that
    have been filtered out throughout for various reasons.
}
\description{
The second function to the word analysis with ginko. This
    function accepts a metaDataMatrix from the \code{\link[ginko]{getScopusMetaData}}
    or the basic processed pdf file data frame of \code{\link[ginko]{createTextMatrixFromPDF}}
    and restructures the data for later use.
}
\examples{
\dontrun{
placeholder}
}
\seealso{
\code{\link{createTextMatrixFromPDF}} for the preceding step,
    \code{\link{calculateModels}} for the proceeding step

Other ginko functions: \code{\link{calculateModels}},
  \code{\link{createOrdinationPlot}},
  \code{\link{createTextMatrixFromPDF}},
  \code{\link{inspectGinko}},
  \code{\link{mostImportantPaperPerCluster}},
  \code{\link{ordinationCluster}}
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
\concept{ginko functions}
