% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus.R
\name{searchScopus}
\alias{searchScopus}
\title{searchScopus}
\usage{
searchScopus(searchString, myAPIKey, maxResults = 10,
  countIncrement = 200)
}
\arguments{
\item{searchString}{The search string you want to ask the server. See the
Scopus API (\url{https://dev.elsevier.com/}) for details.}

\item{myAPIKey}{Your private Elsevier API key for the server communication.
You can request one at \url{https://dev.elsevier.com/}.}

\item{maxResults}{The maximum amount of accepted search results. Usually,
Scopus does not provide more than 5000 results.}

\item{countIncrement}{The number of resuts per GET request. A private user
can't exeed 25 per request. If you are inside an subscribed IP range,
you can use the maximum of 200 per request.}
}
\value{
A data frame containing the DOI numbers and Scopus-IDs of the search
    results, as well as some placeholder columns.
}
\description{
This function accepts an search string in URL format and an
    Elsevier API Key and returns the DOI numbers of all search results as
    a vector for further processing.
}
\examples{
\dontrun{
DOInumbers <- searchScopus('TITLE-ABS-KEY(sustainability) AND PUBYEAR > 2009',
    '1234567890ABCDEF', maxResults = 160, countIncrement = 20)
DOInumbers}
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
